+++ vscode-test-runner/workspace-1/.joyride/src/integration_test/runner.cljs
@@ -40,15 +40,18 @@
       (p/resolve! running true)
       (p/reject! running true))))
 
+;; We rely on that the user_activate.cljs script is run before workspace_activate.cljs
 (defn- run-when-ws-activated [tries]
   (if (:ws-activated? @db/!state)
     (do
       (println "Runner: Workspace activated, running tests")
+      (require '[integration-test.user-activate-test])
       (require '[integration-test.workspace-activate-test])
       (require '[integration-test.ws-scripts-test])
       (require '[integration-test.require-js-test])
       (require '[integration-test.npm-test])
-      (cljs.test/run-tests 'integration-test.workspace-activate-test
+      (cljs.test/run-tests 'integration-test.user-activate-test
+                           'integration-test.workspace-activate-test
                            'integration-test.ws-scripts-test
                            'integration-test.require-js-test
                            'integration-test.npm-test))