+++ vscode-test-runner/workspace-1/.joyride/src/integration_test/user_activate_test.cljs
@@ -1,0 +1,13 @@
+(ns integration-test.user-activate-test
+  (:require [cljs.test :refer [deftest testing is]]
+            ["path" :as path]))
+
+(deftest user-activate 
+  (testing "User activation script is required"
+    (is #_{:clj-kondo/ignore [:unresolved-namespace]}
+     (= #'user-activate/!db
+        ((ns-publics 'user-activate) '!db))))
+  
+  (testing "my-lib is required"
+    (is (seq
+         (ns-publics 'my-lib)))))