+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changes to Joyride
 
 ## [Unreleased]
 
+- [Enable using JS files as user and workspace scripts](https://github.com/BetterThanTomorrow/joyride/issues/132)
 - Fix [Allow `js/require` to be used in joyride](https://github.com/BetterThanTomorrow/joyride/issues/134)
 
 ## [0.0.29] - 2023-01-02