+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to Joyride
 
 - [Enable using JS files as user and workspace scripts](https://github.com/BetterThanTomorrow/joyride/issues/132)
 - Fix [Allow `js/require` to be used in joyride](https://github.com/BetterThanTomorrow/joyride/issues/134)
+- Dev internals [Add basic e2e tests for user scripts](https://github.com/BetterThanTomorrow/joyride/issues/136)
 
 ## [0.0.29] - 2023-01-02