+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changes to Joyride
 
 ## [Unreleased]
 
+## [0.0.30] - 2023-01-06
+
 - [Enable using JS files as user and workspace scripts](https://github.com/BetterThanTomorrow/joyride/issues/132)
 - Fix: [Allow `js/require` to be used in joyride](https://github.com/BetterThanTomorrow/joyride/issues/134)
 - Dev internals: [Add basic e2e tests for user scripts](https://github.com/BetterThanTomorrow/joyride/issues/136)